Refactors connect command.
* Renames csv package to loader.
* Renames cli package to session.
* Groups csv-related functions together.
* Cleans up csv writer instantiation.
