+++ cmd/connect/session/gap.go
@@ -1,4 +1,4 @@
-package cli
+package session
 
 import (
 	"fmt"
