+++ cmd/connect/session/help.go
@@ -1,4 +1,4 @@
-package cli
+package session
 
 import (
 	"fmt"
