+++ cmd/connect/session/sql.go
@@ -1,4 +1,4 @@
-package cli
+package session
 
 import (
 	"fmt"
